- Move hasCSSTOMSupport into core - Improve default unit selection
