+++ flow-typed/css.js
@@ -1,0 +1,8 @@
+interface CSSOM {
+  px(val: number | string): {};
+  percent(val: number | string): {};
+  ms(val: number | string): {};
+  number(val: number | string): {};
+}
+
+declare var CSS: ?CSSOM
