+++ packages/react-jss/.size-snapshot.json
@@ -1,13 +1,13 @@
 {
   "dist/react-jss.js": {
-    "bundled": 125073,
-    "minified": 43593,
-    "gzipped": 13026
+    "bundled": 125335,
+    "minified": 43283,
+    "gzipped": 13002
   },
   "dist/react-jss.min.js": {
-    "bundled": 101665,
-    "minified": 36422,
-    "gzipped": 10932
+    "bundled": 101927,
+    "minified": 36112,
+    "gzipped": 10911
   },
   "dist/react-jss.cjs.js": {
     "bundled": 16755,
